@@ -5,12 +5,12 @@ plugins {
 
 android {
     namespace 'com.example.giftme'
-    compileSdk 32
+    compileSdk 33
 
     defaultConfig {
         applicationId "com.example.giftme"
         minSdk 22
-        targetSdk 32
+        targetSdk 33
         versionCode 1
         versionName "1.0"
 
@@ -33,11 +33,13 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.5.1'
     implementation 'com.google.android.material:material:1.7.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-    implementation 'com.google.firebase:firebase-firestore:24.4.0'
+    implementation 'com.google.firebase:firebase-firestore:24.4.1'
     implementation 'androidx.preference:preference:1.2.0'
+    implementation 'com.squareup.picasso:picasso:2.8'
+
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.4'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
 
     //to solve duplicate class androidx.lifecycle.ViewModeLazy issue
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1'
@@ -50,5 +52,5 @@ dependencies {
     implementation 'com.google.firebase:firebase-auth'
 
     // Also add the dependency for the Google Play services library and specify its version
-    implementation 'com.google.android.gms:play-services-auth:20.3.0'
+    implementation 'com.google.android.gms:play-services-auth:20.4.0'
 }
