@@ -1,13 +1,14 @@
-package com.example.giftme;
+package com.example.giftme.Adapters;
 
 import android.app.Activity;
+import android.app.AlertDialog;
 import android.content.Context;
 import android.content.Intent;
+import android.database.Cursor;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageButton;
 import android.widget.LinearLayout;
 import android.widget.RatingBar;
 import android.widget.TextView;
@@ -15,17 +16,20 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import org.w3c.dom.Text;
+import com.example.giftme.Helpers.DataBaseHelper;
+import com.example.giftme.Activities.DetailedItemViewActivity;
+import com.example.giftme.Helpers.Item;
+import com.example.giftme.R;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class ItemsAdapter extends RecyclerView.Adapter<ItemsAdapter.ViewHolder> {
 
     Context context;
     Activity activity;
-    private TextView collectionNameTV;
+    TextView collectionNameTV;
     List<Item> myItems;
+    DataBaseHelper dataBaseHelper;
 
     public ItemsAdapter(Activity activity, Context context, List<Item> items) {
         this.context = context;
@@ -52,10 +56,8 @@ public class ItemsAdapter extends RecyclerView.Adapter<ItemsAdapter.ViewHolder> 
         itemName.setText(item.getName());
         RatingBar ratingBar = holder.ratingBar;
         ratingBar.setRating(item.getHearts());
-        
         String collectionName = (String) collectionNameTV.getText();
         holder.currentItem = myItems.get(index);
-
         holder.linearLayout.setOnClickListener(view -> {
             Intent intent = new Intent(this.activity, DetailedItemViewActivity.class);
             //put in name, price description, hearts, and link etc
@@ -70,10 +72,48 @@ public class ItemsAdapter extends RecyclerView.Adapter<ItemsAdapter.ViewHolder> 
             intent.putExtra("collectionName", collectionName);
             this.activity.finish();
             this.activity.startActivity(intent);
-
         });
-
+        holder.linearLayout.setOnLongClickListener(view -> {
+            int index2 = holder.getAdapterPosition();
+            confirmDialogForDeleteItem(index2, collectionName);
+            return true;
+        });
     }
+    private void confirmDialogForDeleteItem(int position, String collectionName) {
+
+        AlertDialog.Builder builder = new AlertDialog.Builder(context);
+        builder.setTitle("Delete " + myItems.get(position).getName() + " ?");
+        builder.setMessage("Items details will also be deleted!");
+        builder.setPositiveButton("Yes", (dialogInterface, i) -> {
+            dataBaseHelper = new DataBaseHelper(context);
+            dataBaseHelper.deleteItemInCollection(String.valueOf(myItems.get(position).getId()),collectionName);
+            myItems.clear();
+            getAllItems();
+            notifyItemRemoved(position);
+        });
+        builder.setNegativeButton("No", (dialogInterface, i) -> dialogInterface.cancel());
+        builder.create().show();
+    }
+    private void getAllItems() {
+        Cursor cursor = dataBaseHelper.selectAll(collectionNameTV.getText().toString());
+        if (cursor.getCount() != 0) {
+            while (cursor.moveToNext()) {
+                Item currentItem
+                        = new Item(Integer.parseInt(cursor.getString(0)),
+                        cursor.getString(1),
+                        cursor.getString(2),
+                        Integer.parseInt(cursor.getString(3)),
+                        Integer.parseInt(cursor.getString(4)),
+                        cursor.getString(5),
+                        cursor.getString(6),
+                        cursor.getString(7)
+                );
+                myItems.add(currentItem);
+            }
+            cursor.close();
+        }
+    }
+
 
     @Override
     public int getItemCount() {
@@ -90,10 +130,8 @@ public class ItemsAdapter extends RecyclerView.Adapter<ItemsAdapter.ViewHolder> 
 
         public ViewHolder(@NonNull View itemView) {
             super(itemView);
-            
             itemNameTV = itemView.findViewById(R.id.item_name);
             ratingBar = itemView.findViewById(R.id.rating);
-
             linearLayout = itemView.findViewById(R.id.item_lv);
         }
     }
