@@ -1,4 +1,4 @@
-package com.example.giftme;
+package com.example.giftme.Adapters;
 
 import android.app.Activity;
 import android.content.Context;
@@ -9,11 +9,14 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.LinearLayout;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.giftme.Helpers.DataBaseHelper;
+import com.example.giftme.Activities.MyCollectionItems;
+import com.example.giftme.R;
 
 import java.util.ArrayList;
 
