@@ -1,7 +1,8 @@
-package com.example.giftme;
+package com.example.giftme.Fragments;
 
 import android.content.Context;
 import android.content.Intent;
+import android.database.Cursor;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
@@ -12,6 +13,11 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.giftme.Adapters.MyCollectionItemsAdapter;
+import com.example.giftme.Activities.AddNewItemManually;
+import com.example.giftme.Helpers.DataBaseHelper;
+import com.example.giftme.Helpers.Item;
+import com.example.giftme.R;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import java.util.ArrayList;
@@ -24,8 +30,7 @@ public class DetailViewFragment extends Fragment {
     FloatingActionButton actionButton;
     MyCollectionItemsAdapter myCollectionItemsAdapter;
     DataBaseHelper dataBaseHelper;
-    ArrayList<String> name, price, imagePath, date;
-    ArrayList<Integer> favorite;
+    ArrayList<Item> items;
     String collection_name;
 
     public DetailViewFragment() {
@@ -41,6 +46,7 @@ public class DetailViewFragment extends Fragment {
         recyclerView = view.findViewById(R.id.recycler_view_detail_view);
         actionButton = view.findViewById(R.id.detailed_view_action);
         dataBaseHelper = new DataBaseHelper(context);
+        items = new ArrayList<>();
         if (getArguments() != null) {
             collection_name = getArguments().getString("collection_name");
         }
@@ -49,19 +55,31 @@ public class DetailViewFragment extends Fragment {
             intent.putExtra("collection_name", collection_name);
             startActivity(intent);
         });
-        name = new ArrayList<>();
-        name.add("iPhone 14 Pro Max");
-        price = new ArrayList<>();
-        price.add("$999");
-        imagePath = new ArrayList<>();
-        date = new ArrayList<>();
-        date.add("2022-11-20");
-        favorite = new ArrayList<>();
-        myCollectionItemsAdapter = new MyCollectionItemsAdapter(getActivity(), context, name,
-                price, imagePath, favorite, date);
+        getAllItems();
+        myCollectionItemsAdapter = new MyCollectionItemsAdapter(getActivity(), context, items);
         recyclerView.setLayoutManager(new LinearLayoutManager(context));
         recyclerView.setHasFixedSize(true);
         recyclerView.setAdapter(myCollectionItemsAdapter);
         return view;
+    }
+
+    private void getAllItems() {
+        Cursor cursor = dataBaseHelper.selectAll(collection_name);
+        if (cursor.getCount() != 0) {
+            while (cursor.moveToNext()) {
+                Item currentItem
+                        = new Item(Integer.parseInt(cursor.getString(0)),
+                        cursor.getString(1),
+                        cursor.getString(2),
+                        Integer.parseInt(cursor.getString(3)),
+                        Integer.parseInt(cursor.getString(4)),
+                        cursor.getString(5),
+                        cursor.getString(6),
+                        cursor.getString(7)
+                );
+                items.add(currentItem);
+            }
+            cursor.close();
+        }
     }
 }
