@@ -1,4 +1,6 @@
-package com.example.giftme;
+package com.example.giftme.Helpers;
+
+import androidx.annotation.NonNull;
 
 import android.util.Log;
 
@@ -9,11 +11,12 @@ public class Item {
 
     private int id;
     private String name;
-    private int hearts;
+    private float hearts;
     private int price;
+    private String website;
     private String description;
     private String date;
-    private int img; //may change
+    private String img;
     private String tableID; //FIRESTORE_ID
 
     public Item(){
@@ -21,8 +24,9 @@ public class Item {
         // or if it is being created for the database
         setTableID();
     }
-    public Item(int newId, String newName, int newHearts, int newPrice, String newDescription, String newDate, int newImg){
-        setId(newId);
+    public Item(int newId, String website, String newName, float newHearts, int newPrice, String newDescription, String newDate, String newImg){
+        setWebsite(website);
+        setId( newId);
         setName(newName);
         setHearts(newHearts);
         setPrice(newPrice);
@@ -34,29 +38,31 @@ public class Item {
 
     //setters
     public void setId(int newId){ id = newId; }
+    public void setWebsite(String URL) {website = URL;}
     public void setName( String newName){ name = newName; }
-    public void setHearts(int newHearts){ hearts = newHearts;}
+    public void setHearts(float newHearts){ hearts = newHearts;}
     public void setPrice(int newPrice){price = newPrice;}
     public void setDescription( String newDescription){description = newDescription;}
-    public void setImg( int newImg){ img = newImg;}
+    public void setImg( String newImg){ img = newImg;}
     public void setDate(String newDate){ date = newDate;}
     // using Java's UUID class to generate a unique ID for each item (cryptographically strong pseudo random number generator)
     public void setTableID(){ tableID = UUID.randomUUID().toString();}
 
     //getters
     public int getId() {return id;}
+    public String getWebsite() {return website;}
     public String getName() {return name;}
-    public int getHearts() {return hearts;}
+    public float getHearts() {return hearts;}
     public int getPrice() {return price;}
     public String getDescription() {return description;}
     public String getDate(){return date;}
-    public int getImg(){return img;}
+    public String getImg(){return img;}
     public String getTableID() {return tableID;}
 
     //toString
+    @NonNull
     @Override
     public String toString(){
         return id +"; " + name +"; " + hearts +"; " +price + "; " + description + "; " + date +"; " + img + "; " + tableID + "; " ;
     }
-
 }
